requirements.txt
fastapi
uvicorn
faker
sse-starlette
pymongo
mysql-connector-python